55 add sending images

+++ client/src/services/uploadUserImage.js
@@ -6,15 +6,23 @@ let storageRef = firebase.storage().ref();
 const replaceUserImage = async (newImageData, oldImageData, userEmail) => {
   let desertRef;
   try {
-    await axios.put(`user/${userEmail}/image`, {
-      newImageData: { ...newImageData },
-    });
-    if (oldImageData !== undefined && !!oldImageData.name) {
+    if (oldImageData !== null) {
+      await axios.put(`user/${userEmail}/image`, {
+        newImageData: { ...newImageData },
+      });
+    } else {
+      console.log("USER PICTURE FOR MESSAGE FIELD UPLOADED TO FIREBASE...");
+    }
+    if (
+      oldImageData !== null &&
+      oldImageData !== undefined &&
+      !!oldImageData.name
+    ) {
       desertRef = storageRef.child(`photos/${oldImageData.name}`);
       await desertRef.delete();
     }
   } catch (err) {
-    throw Error("Sorry something went wrong ", err.message);
+    throw Error("Oops, something went wrong ", err.message);
   }
 };
 
@@ -27,7 +35,15 @@ const uploadUserImage = async (newImage, oldImageData, userEmail) => {
       writable: true,
     },
   });
-  const fileRef = storageRef.child(`photos/${newImage.name}`);
+
+  let fileRef;
+
+  if (oldImageData === null) {
+    fileRef = storageRef.child(`user-images/${newImage.name}`);
+  } else {
+    fileRef = storageRef.child(`photos/${newImage.name}`);
+  }
+
   await fileRef.put(newImage, { name: newImage.name });
   const newImageUrl = await fileRef.getDownloadURL();
   const newImageData = {
@@ -35,6 +51,7 @@ const uploadUserImage = async (newImage, oldImageData, userEmail) => {
     name: newImage.name,
   };
   await replaceUserImage(newImageData, oldImageData, userEmail);
+  return newImageData;
 };
 
 export default uploadUserImage;
